Allow creating self-managed consumers.
The consumer application need not use Kafka's built-in offset storage, it can store offsets in a store of its own choosing. The primary use case for this is allowing the application to store both the offset and the results of the consumption in the same system in a way that both the results and offsets are stored atomically. This is not always possible, but when it is it will make the consumption fully atomic and give "exactly once" semantics that are stronger than the default "at-least once" semantics you get with Kafka's offset commit functionality.

To manage your own offset you just need to do the following:

    Configure enable.auto.commit=false
    Use the offset provided with each ConsumerRecord to save your position.
    On restart restore the position of the consumer using seek(TopicPartition, long).